@grosser 2021-01-16
looks fine to me, lmk if you are done and I'll merge, then we can have a look at perf improvements / smaller details 👍